+++ spec/rubocop/cop/performance/constant_regexp_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe RuboCop::Cop::Performance::ConstantRegexp do
   it 'registers an offense and corrects when regexp contains interpolated constant' do
     expect_offense(<<~RUBY)
       str.match?(/\A\#{CONST}/)
-                 ^^^^^^^^^^^ Extract this regexp into a constant or append an `/o` option to its options.
+                 ^^^^^^^^^^^ Extract this regexp into a constant, memoize it, or append an `/o` option to its options.
     RUBY
 
     expect_correction(<<~RUBY)
@@ -17,7 +17,7 @@ RSpec.describe RuboCop::Cop::Performance::ConstantRegexp do
   it 'registers an offense and corrects when regexp contains multiple interpolated constants' do
     expect_offense(<<~RUBY)
       str.match?(/\A\#{CONST1}something\#{CONST2}\z/)
-                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Extract this regexp into a constant or append an `/o` option to its options.
+                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Extract this regexp into a constant, memoize it, or append an `/o` option to its options.
     RUBY
 
     expect_correction(<<~RUBY)
@@ -28,7 +28,7 @@ RSpec.describe RuboCop::Cop::Performance::ConstantRegexp do
   it 'registers an offense and corrects when regexp contains `Regexp.escape` on constant' do
     expect_offense(<<~RUBY)
       str.match?(/\A\#{CONST1}something\#{Regexp.escape(CONST2)}\z/)
-                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Extract this regexp into a constant or append an `/o` option to its options.
+                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Extract this regexp into a constant, memoize it, or append an `/o` option to its options.
     RUBY
 
     expect_correction(<<~RUBY)
@@ -45,6 +45,12 @@ RSpec.describe RuboCop::Cop::Performance::ConstantRegexp do
   it 'does not register an offense when regexp is within assignment to a constant' do
     expect_no_offenses(<<~RUBY)
       CONST = str.match?(/\#{ANOTHER_CONST}/)
+    RUBY
+  end
+
+  it 'does not register an offense when the regexp is memoized' do
+    expect_no_offenses(<<~RUBY)
+      @regexp ||= /\#{ANOTHER_CONST}/
     RUBY
   end